+++ dev/client/main.go
@@ -27,9 +27,9 @@ func main() {
 
 	privKey := []string{
 		"8ff3ca2d9985c3a52b459e2f6e7822b23e1af845961e22128d5f372fb9aa5f17",
-		"171786c73f805d257ceb07206d851eea30b3b41a2170ae55e1225e0ad516ef42",
-		"b7700998b973a2cae0cb8e8a328171399c043e57289735aca5f2419bd622297a",
-		"00dcf944648491b3a822d40bf212f359f699ed0dd5ce5a60f1da5e1142855949",
+		//"171786c73f805d257ceb07206d851eea30b3b41a2170ae55e1225e0ad516ef42",
+		//"b7700998b973a2cae0cb8e8a328171399c043e57289735aca5f2419bd622297a",
+		//"00dcf944648491b3a822d40bf212f359f699ed0dd5ce5a60f1da5e1142855949",
 	}
 
 	var testFunc func(privKey string, blockTime time.Duration) error
